medical-ai-assistant-github/requirements.txt
langchain>=1.1.0
langchain-community>=0.2.0
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.8.0
chromadb>=0.4.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
jieba>=0.42.1
pypdf>=3.17.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
tiktoken>=0.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
